requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
python-dotenv>=1.0.0
